+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   - Support completion on aliased keywords. #649
   - Add new `Sort map keys` refactoring code action. #651
   - Add new `Create function` code action, allowing to create a function on a existing namespace or creating a new namespace + the function. #646
+  - Improve `Extract function` refactoring to consider comments above current function. 
 
 ## 2021.11.16-16.52.14